+++ src/client/gui/dialogs/client_dialogdataviewitem.h
@@ -38,45 +38,12 @@ public:
     explicit mbClientDialogDataViewItem(QWidget *parent = nullptr);
     ~mbClientDialogDataViewItem();
 
-public:
-    MBSETTINGS getSettings(const MBSETTINGS &settings = MBSETTINGS(), const QString &title = QString()) override;
-
 private:
-    void fillForm(const MBSETTINGS &settings);
-    void fillFormByteOrder(mb::DataOrder e);
-    void fillFormRegisterOrder(mb::DataOrder e, mbClientDevice *dev = nullptr);
-    void fillFormByteArrayFormat(mb::DigitalFormat e, mbClientDevice *dev = nullptr);
-    void fillFormByteArraySeparator(const QString &s, mbClientDevice *dev = nullptr);
-    void fillFormStringLengthType(mb::StringLengthType e, mbClientDevice *dev = nullptr);
-    void fillFormStringEncoding(mb::StringEncoding e, mbClientDevice *dev = nullptr);
-
-    void fillData(MBSETTINGS &settings);
-    void fillDataByteOrder(MBSETTINGS &settings);
-    void fillDataRegisterOrder(MBSETTINGS &settings);
-    void fillDataByteArrayFormat(MBSETTINGS &settings);
-    void fillDataByteArraySeparator(MBSETTINGS &settings);
-    void fillDataStringLengthType(MBSETTINGS &settings);
-    void fillDataStringEncoding(MBSETTINGS &settings);
-
-private Q_SLOTS:
-    void deviceChanged(int i);
-    void adrTypeChanged(int i);
-    void formatChanged(int i);
-    void lengthChanged(int i);
-    void togleDefaultByteArraySeparator();
-
-private:
-    void setVariableLength(int len);
-    bool isDefaultByteArraySeparator();
-    inline QString nonDefaultByteArraySeparator() const { return m_nonDefaultByteArraySeparator; }
-    void setNonDefaultByteArraySeparator(const QString &s);
+    void fillFormInner(const MBSETTINGS &settings) override;
+    void fillDataInner(MBSETTINGS &settings) override;
 
 private:
     Ui::mbClientDialogDataViewItem *ui;
-
-    mb::Format m_formatLast;
-    int m_variableLength;
-    QString m_nonDefaultByteArraySeparator;
 };
 
 #endif // CLIENT_DIALOGDATAVIEWITEM_H