+++ changelog.md
@@ -5,5 +5,10 @@
 
 # 0.1.2
 
-* Tested and added binaries for Ubuntu 22.04. 
 * Fixed some bugs
+
+# 0.1.3
+
+* Enable Server side 'Edit'-menu commands (Paste,Insert,Delete) for DataView table in 'Run' mode
+* Improve import of DataView items
+* Fix some bugs